added main monitoring loop
copied the primary monitoring loop from test program. in every loop it will check powerbank and reactor status and turn the reactor on or off accordingly. it also waits a second every loop and iterates a counter that is used in measuring how long it took for the power bank to be either filled or emtied